Rename and simplify about/experience sections
